VolumeDialog: notifyVisible after the volume panel is fully hidden
notifyVisible(false) leads to reset of active stream. As a result,
the current active row will be untinted. Currently, this happens
immediately after the dismissal is requested, before the volume panel
is fully hidden.

User can notice the row has been untinted before the panel is hidden.

This changes moves notifyVisible to the end of animation so the volume
panel appearance will be consistent in between.

Change-Id: Iacd83d48634a11bc248c9a2a45b68b53256fdcb2